Eng 582 support search in dynamic adapters (#1877)
# Support SEARCH in dynamic adapters

We currently have all the necessary meta data for creating a searchable
adapter.
We only need to:

* Create a searchable adapters
* Initiate them during source startup

It is important to note that, in the SEARCH method we need to check the
query to see if it is actually for a terraform mappings and targeting a
single resource rather than multiple. Simply, if the query contains and
`/`, then it is for a single query, we should use the
`externalCallSingle` and return the single response within a slice of
`items` to comply with the `Search` method signature. Relevant:
[ENG-580](https://linear.app/overmind/issue/ENG-580/handle-terraform-mappings-in-search-method)

GitOrigin-RevId: 13ee3ef6b5e51acc20b787ff9a728523d5a2e96c