fix: ensure to use the GET query param as the unique attribute value in gcp manual adapters (#1976)
# Review manual gcp adapters to ensure the unique attribute is the same
as the query string

Linking items requires using the same string for the `get` query
parameters and the `uniqueAttr` value.
So, we need to make sure the unique attribute value is exactly the same
the query attribute.
For example we need to fix the `iam-service-account-key`.
Currently: the `name` is used as unique attribute which is
`projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}/keys/{key}`. But it
should be in the format of `{seavice_account_name}|{key_name}`. Because
this what the expected `get` query.

## Metadata
- URL:
[https://linear.app/overmind/issue/ENG-553/review-manual-gcp-adapters-to-ensure-the-unique-attribute-is-the-same](https://linear.app/overmind/issue/ENG-553/review-manual-gcp-adapters-to-ensure-the-unique-attribute-is-the-same)
- Identifier: ENG-553
- Status: In Review
- Priority: High
- Assignee: omer.demirok@overmind.tech
- Labels: Bug, Gcp-source
- Project: [GCP
Support](https://linear.app/overmind/project/gcp-support-871ebbdcd5f9).
- Project milestone: Box.com first demo
- Created: 2025-06-06T19:23:10.953Z
- Updated: 2025-06-23T11:07:49.065Z

## Pull requests

- [#1774 Review manual gcp adapters to ensure the unique attribute is
the same as the query
string](https://github.com/overmindtech/workspace/issues/1774)
- [fix: ensure to use the GET query param as the unique attribute value
in gcp manual
adapters](https://github.com/overmindtech/workspace/pull/1976)

## Comments

- GitHub integration:

This comment thread is synced to a corresponding [GitHub
issue](https://github.com/overmindtech/workspace/issues/1774). All
replies are displayed in both locations.

GitOrigin-RevId: c3d8f8de9d612ddacbf9a7c235134df8f6b6de7a