check get project permission during gcp source startup and in heartbeat (#2361)
# Mark GCP Source Health based on minimum permission on project

When customer pushes the `Create Source` button, we immediately create a
GCP source and mark it as `healthy`. This is the case even for a made-up
project name. So, in the GCP source, we should validate the whether we
can get the metadata for the customer project. This will confirm that
the validity of the project and customer actually took an action and
gave the required permissions (not necessarily all, but at least they
did the bare minimum). We can do this check as a health check.

In the UI, if the source is `unhealthy` then we should show the
instructions for how to give permissions to the Overmind SA. Otherwise,
source is healthy, we can collapse/hide the instructions and just show
config for the source.

GitOrigin-RevId: 3a7a7698c68dfd66e1d69bf170539595c99efc3c