Fix LLM Blast Radius to correctly query and wait for recursive results (#2545)
A "proper" fix would require adding a ParentQueryUUID field to sdp.Query
and patching this through all query evaluations to link recursive
queries and their parents. This sounds like more effort than is worth it
here. Instead this creates a separate sdpws.Client for each query and
captures all results of that one connection. To that end, `WaitAll`,
which was not used anywhere, also needed improving to wait for results
from queries that were started (through recursion) during waiting.

This also contains a change to add links to affected items to make the
final blast radius snapshot more useful.

GitOrigin-RevId: 557c2336c12e83d2a61f46eaa6b3ac1fe82c363d